Moved 'EbolaRiskList' component to it's own directory
