+++ src/components/EbolaRiskList/index.js
